Move random test generation to its own file
